Move Negative controls to dedicated right panel
