+++ src/components/panel/right/ControlsPanel.tsx
@@ -7,7 +7,6 @@ import CurveGraph from '../../adjustments/Curves';
 import ColorPanel from '../../adjustments/Color';
 import DetailsPanel from '../../adjustments/Details';
 import EffectsPanel from '../../adjustments/Effects';
-import NegativePanel from '../../adjustments/Negative';
 import CollapsibleSection from '../../ui/CollapsibleSection';
 import Waveform from '../editor/Waveform';
 import Resizer from '../../ui/Resizer';
@@ -278,7 +277,6 @@ export default function Controls() {
             color: ColorPanel,
             details: DetailsPanel,
             effects: EffectsPanel,
-            negative: NegativePanel,
           }[sectionName];
 
           const title = sectionName.charAt(0).toUpperCase() + sectionName.slice(1);
