+++ src/components/ui/AppProperties.tsx
@@ -125,6 +125,7 @@ export enum Panel {
   Export = 'export',
   Masks = 'masks',
   Metadata = 'metadata',
+  Negative = 'negative',
   Presets = 'presets',
 }
