+++ src/store/useUIStore.ts
@@ -7,7 +7,6 @@ export interface CollapsibleSectionsState {
   curves: boolean;
   details: boolean;
   effects: boolean;
-  negative: boolean;
 }
 
 export interface ConfirmModalState {
@@ -138,7 +137,7 @@ export const useUIStore = create<UIState>((set, get) => ({
   activeRightPanel: Panel.Adjustments,
   renderedRightPanel: Panel.Adjustments,
   slideDirection: 1,
-  collapsibleSectionsState: { basic: true, color: false, curves: true, details: false, effects: false, negative: false },
+  collapsibleSectionsState: { basic: true, color: false, curves: true, details: false, effects: false },
 
   isCreateFolderModalOpen: false,
   isRenameFolderModalOpen: false,
